# pkg/gateway/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lenny Gateway API",
    "description": "Lenny session, blob, and admin endpoints. See spec/15_external-api-surface.md.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "/", "description": "Gateway root" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "§10.2 OIDC/OAuth 2.1 access token."
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/credential-pools/{name}/credentials/{credId}": {
      "put": {
        "operationId": "putV1AdminCredentialPoolsByNameCredentialsByCredId",
        "summary": "Update a credential in a pool",
        "x-lenny-mcp-tool": "admin.update_pool_credential",
        "x-lenny-scope": "admin.credential-pools.write",
        "x-lenny-required-role": "tenant-admin",
        "x-lenny-category": "policy-management",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "credId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "operationId": "deleteV1AdminCredentialPoolsByNameCredentialsByCredId",
        "summary": "Remove a credential from a pool",
        "x-lenny-mcp-tool": "admin.remove_pool_credential",
        "x-lenny-scope": "admin.credential-pools.write",
        "x-lenny-required-role": "tenant-admin",
        "x-lenny-category": "policy-management",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "credId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/credentials": {
      "post": {
        "operationId": "postV1Credentials",
        "summary": "Register a credential for the authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      },
      "get": {
        "operationId": "getV1Credentials",
        "summary": "List the authenticated user's registered credentials",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/credentials/{credentialRef}": {
      "put": {
        "operationId": "putV1CredentialsByCredentialRef",
        "summary": "Rotate (replace) secret material for an existing credential",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "credentialRef", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "operationId": "deleteV1CredentialsByCredentialRef",
        "summary": "Remove a registered credential",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "credentialRef", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/v1/credentials/{credentialRef}/revoke": {
      "post": {
        "operationId": "postV1CredentialsByCredentialRefRevoke",
        "summary": "Revoke a credential and invalidate all active leases backed by it",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "credentialRef", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}